refractor: delete modal
