+++ public/pages/Aliases/containers/AliasActions/index.tsx
@@ -2,15 +2,11 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-import React, { useCallback, useContext, useMemo, useState } from "react";
+import React, { useMemo, useState } from "react";
 import { EuiButton, EuiContextMenu } from "@elastic/eui";
 
 import SimplePopover from "../../../../components/SimplePopover";
-import { CoreServicesContext } from "../../../../components/core_services";
-import DeleteIndexModal from "../../components/DeleteAliasModal";
-import { ServicesContext } from "../../../../services";
-import { BrowserServices } from "../../../../models/interfaces";
-import { CoreStart } from "opensearch-dashboards/public";
+import DeleteIndexModal from "../DeleteAliasModal";
 import { IAlias } from "../../interface";
 
 export interface IndicesActionsProps {
@@ -21,30 +17,10 @@ export interface IndicesActionsProps {
 export default function IndicesActions(props: IndicesActionsProps) {
   const { selectedItems, onDelete } = props;
   const [deleteIndexModalVisible, setDeleteIndexModalVisible] = useState(false);
-  const coreServices = useContext(CoreServicesContext) as CoreStart;
-  const services = useContext(ServicesContext) as BrowserServices;
 
   const onDeleteIndexModalClose = () => {
     setDeleteIndexModalVisible(false);
   };
-
-  const onDeleteIndexModalConfirm = useCallback(async () => {
-    const indexPayload = selectedItems.map((item) => item.alias);
-    const result = await services.commonService.apiCaller({
-      endpoint: "indices.deleteAlias",
-      data: {
-        index: "_all",
-        name: selectedItems.map((item) => item.alias),
-      },
-    });
-    if (result && result.ok) {
-      coreServices.notifications.toasts.addSuccess(`Delete [${indexPayload.join(",")}] successfully`);
-      setDeleteIndexModalVisible(false);
-      onDelete();
-    } else {
-      coreServices.notifications.toasts.addDanger(result?.error || "");
-    }
-  }, [selectedItems, services, coreServices, onDelete, onDeleteIndexModalClose]);
 
   const renderKey = useMemo(() => Date.now(), [selectedItems]);
 
@@ -80,10 +56,13 @@ export default function IndicesActions(props: IndicesActionsProps) {
         />
       </SimplePopover>
       <DeleteIndexModal
-        selectedItems={selectedItems.map((item) => item.index)}
+        selectedItems={selectedItems.map((item) => item.alias)}
         visible={deleteIndexModalVisible}
         onClose={onDeleteIndexModalClose}
-        onConfirm={onDeleteIndexModalConfirm}
+        onDelete={() => {
+          onDeleteIndexModalClose();
+          onDelete();
+        }}
       />
     </>
   );
