+++ public/pages/Aliases/containers/DeleteAliasModal/index.ts
